fix(proxy): move stream transport resolution to support domain
The http_bridge domain imported _resolved_configured_stream_transport
from the streaming domain, violating the proxy service architecture
ratchet (http_bridge may not import streaming). Move the helper into
the shared support domain, which both http_bridge and streaming are
allowed to import.
